Move validator stats call in eth2 module
